moving fixtures to fixture package
