+++ src/test/java/net/vidageek/mirror/dsl/MirrorTest.java
@@ -9,9 +9,9 @@ import java.lang.reflect.Method;
 
 import net.vidageek.mirror.exception.MirrorException;
 import net.vidageek.mirror.fixtures.ClassFixture;
-import net.vidageek.mirror.proxy.OneClass;
-import net.vidageek.mirror.proxy.OneInterface;
-import net.vidageek.mirror.proxy.OtherClass;
+import net.vidageek.mirror.fixtures.OneClassFixture;
+import net.vidageek.mirror.fixtures.OneInterfaceFixture;
+import net.vidageek.mirror.fixtures.OtherClassFixture;
 import net.vidageek.mirror.proxy.dsl.MethodInterceptor;
 
 import org.junit.Test;
@@ -75,7 +75,7 @@ public class MirrorTest {
 
 	@Test
 	public void testThatProxifyOneInterfaceIsCorrect() {
-		OneInterface proxy = new Mirror().proxify(OneInterface.class).interceptingWith(new MethodInterceptor() {
+		OneInterfaceFixture proxy = new Mirror().proxify(OneInterfaceFixture.class).interceptingWith(new MethodInterceptor() {
 
 			public boolean accepts(final Method method) {
 				return true;
@@ -91,7 +91,7 @@ public class MirrorTest {
 
 	@Test
 	public void testThatProxifyMoreThanOneInterfaceIsCorrect() {
-		Object proxy = new Mirror().proxify(OneClass.class).interceptingWith(new MethodInterceptor() {
+		Object proxy = new Mirror().proxify(OneClassFixture.class).interceptingWith(new MethodInterceptor() {
 
 			public boolean accepts(final Method method) {
 				return true;
@@ -104,6 +104,6 @@ public class MirrorTest {
 
 		// assertEquals("foo", ((OneInterface) proxy).interfaceMethod());
 		// assertEquals("foo", ((OtherInterface) proxy).otherInterfaceMethod());
-		assertEquals("foo", ((OtherClass) proxy).classMethod());
+		assertEquals("foo", ((OtherClassFixture) proxy).classMethod());
 	}
 }
