+++ src/test/java/net/vidageek/mirror/fixtures/OneInterfaceFixture.java
@@ -1,0 +1,11 @@
+package net.vidageek.mirror.fixtures;
+
+/**
+ * 
+ * @author Juliano Alves
+ *
+ */
+public interface OneInterfaceFixture {
+
+	String interfaceMethod();
+}
