+++ src/test/java/net/vidageek/mirror/fixtures/OtherClassFixture.java
@@ -1,0 +1,8 @@
+package net.vidageek.mirror.fixtures;
+
+public class OtherClassFixture {
+
+	public String classMethod() {
+		return "test";
+	}
+}
