+++ src/test/java/net/vidageek/mirror/fixtures/OtherInterfaceFixture.java
@@ -1,0 +1,6 @@
+package net.vidageek.mirror.fixtures;
+
+public interface OtherInterfaceFixture {
+
+	int otherInterfaceMethod();
+}
