--- src/test/java/net/vidageek/mirror/proxy/OneInterface.java
@@ -1,11 +1,0 @@
-package net.vidageek.mirror.proxy;
-
-/**
- * 
- * @author Juliano Alves
- *
- */
-public interface OneInterface {
-
-	String interfaceMethod();
-}
