--- src/test/java/net/vidageek/mirror/proxy/OtherClass.java
@@ -1,8 +1,0 @@
-package net.vidageek.mirror.proxy;
-
-public class OtherClass {
-
-	public String classMethod() {
-		return "test";
-	}
-}
