--- src/test/java/net/vidageek/mirror/proxy/OtherInterface.java
@@ -1,6 +1,0 @@
-package net.vidageek.mirror.proxy;
-
-public interface OtherInterface {
-
-	int otherInterfaceMethod();
-}
